命令模式-画笔实战

+++ app/src/main/java/com/orangeaterz/designpatterns/命令模式/画笔实战/CircleBrush.java
@@ -4,17 +4,17 @@ import android.graphics.Path;
 
 public class CircleBrush implements IBrush {
     @Override
-    public void down(Path path, int x, int y) {
+    public void down(Path path, float x, float y) {
 
     }
 
     @Override
-    public void move(Path path, int x, int y) {
+    public void move(Path path, float x, float y) {
         path.addCircle(x, y, 10, Path.Direction.CW);
     }
 
     @Override
-    public void up(Path path, int x, int y) {
+    public void up(Path path, float x, float y) {
 
     }
 }
