+++ app/src/main/java/com/orangeaterz/designpatterns/命令模式/画笔实战/NormalBrush.java
@@ -4,17 +4,17 @@ import android.graphics.Path;
 
 public class NormalBrush implements IBrush {
     @Override
-    public void down(Path path, int x, int y) {
+    public void down(Path path, float x, float y) {
         path.moveTo(x, y);
     }
 
     @Override
-    public void move(Path path, int x, int y) {
+    public void move(Path path, float x, float y) {
         path.lineTo(x, y);
     }
 
     @Override
-    public void up(Path path, int x, int y) {
+    public void up(Path path, float x, float y) {
 
     }
 }
